# src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MAW API",
    "description": "Multi-Agent Workspace — fleet orchestration, monitoring, and communication API.\n\nAuthentication: Federation HMAC (X-Maw-Signature + X-Maw-Timestamp) on protected endpoints. PIN/JWT for web UI. Loopback requests bypass auth.",
    "version": "1.0.0",
    "contact": { "name": "MAW Team" }
  },
  "servers": [
    { "url": "/api", "description": "Local MAW instance" }
  ],
...
  "components": {
    "securitySchemes": {
      "federationHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Maw-Signature",
        "description": "HMAC-SHA256 signature: sign(token, method, path, timestamp). Also requires X-Maw-Timestamp header."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Tenant API key (maw_<hex>). Get one via POST /api/tenant/keys."
      },
      "pinJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/pin-verify"
      }
    },
    "schemas": {
...
      "Ok": {
        "type": "object",
        "properties": { "ok": { "type": "boolean", "const": true } }
      }
    }
  },
  "paths": {
...
    "/tenant/keys": {
      "post": {
        "tags": ["Tenant"],
        "summary": "Create API key for a trial",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["trialId"], "properties": { "trialId": { "type": "string" }, "name": { "type": "string", "default": "default" } } } } } },
        "responses": { "201": { "description": "API key created", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "apiKey": { "type": "object", "properties": { "id": { "type": "string" }, "key": { "type": "string" }, "name": { "type": "string" }, "tier": { "type": "string" }, "rateLimit": { "type": "integer" } } } } } } } } }
      }
    },
    "/tenant/keys/{trialId}": {
      "get": {
        "tags": ["Tenant"],
        "summary": "List API keys (key preview only)",
        "parameters": [{ "name": "trialId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Key list", "content": { "application/json": { "schema": { "type": "object", "properties": { "keys": { "type": "array", "items": { "type": "object" } }, "total": { "type": "integer" } } } } } } }
      }
    },
    "/tenant/keys/{keyId}": {
      "delete": {
        "tags": ["Tenant"],
        "summary": "Revoke API key",
        "parameters": [{ "name": "keyId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Revoked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
    "/tenant/usage/{trialId}": {
      "get": {
        "tags": ["Tenant"],
        "summary": "Usage stats per tenant",
        "parameters": [
          { "name": "trialId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "hours", "in": "query", "schema": { "type": "integer", "default": 24 } }
        ],
        "responses": { "200": { "description": "Usage breakdown", "content": { "application/json": { "schema": { "type": "object", "properties": { "trialId": { "type": "string" }, "hours": { "type": "integer" }, "totalRequests": { "type": "integer" }, "endpoints": { "type": "array", "items": { "type": "object" } }, "keys": { "type": "integer" }, "tier": { "type": "string" } } } } } } }
      }
    }
  }
}